+++ changelog.md
@@ -61,3 +61,12 @@ ARS \[handlers\] (2020-09-30)
     * Bug Fix
   \[remedy_generic_query_retrieve_v4\] (2020-09-30)
     * Bug Fix
+	
+ARS \[handlers\] (2021-01-06)
+  \[ars9_generic_query_retrieve_v1\] (2021-01-06)
+    * Updated escape method.
+
+ARS \[bridge-adapter\] (2021-02-25)
+  \[kinetic-bridgehub-adapter-ars\]
+    * PER-204 fixed issue that caused field to large for sort error to be thrown.
+    * versioned adapter to v1.0.4